tire-price/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
psycopg2-binary>=2.9.9
playwright>=1.40.0
gspread>=6.0.0
google-auth>=2.25.0
python-dotenv>=1.0.0
